refactor(orm): reduce lines of code
use where to implement find and find_by
add query_helper to decongest and extend query_builder
remove redundant attr_assessors
make unneeded methods private
rename variables, make tests more descriptive

[#123405751]
